% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldruc.R
\name{des}
\alias{des}
\title{Describe the data and display the varible name,label,type and position information in dataview.}
\usage{
des(dfile)
}
\arguments{
\item{dfile}{a dataframe.}
}
\description{
Describe the data and display the varible name,label,type and position information in dataview.
}
\examples{
library(ldtool)
des(mtcars)
}
