% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldruc.R
\name{fre}
\alias{fre}
\title{make frequency table using expss library}
\usage{
fre(
  x,
  weight = NULL,
  drop_unused_labels = TRUE,
  prepend_var_lab = FALSE,
  stat_lab = getOption("expss.fre_stat_lab", c("Count", "Valid percent", "Percent",
    "Responses, \%", "Cumulative responses, \%"))
)
}
\description{
使用expss::fre制作频数表
}
