% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldruc.R
\name{freq}
\alias{freq}
\title{make frequency table using sjmisc library.}
\usage{
freq(
  x,
  ...,
  sort.frq = c("none", "asc", "desc"),
  weights = NULL,
  auto.grp = NULL,
  show.strings = TRUE,
  show.na = TRUE,
  grp.strings = NULL,
  min.frq = 0,
  out = c("txt", "viewer", "browser"),
  title = NULL,
  encoding = "UTF-8",
  file = NULL
)
}
\description{
用来制作频数表，使用sjmisc::frq函数
}
