% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldruc.R
\name{mrtab}
\alias{mrtab}
\title{Frequence tabll of Multiple response variable .}
\usage{
mrtab(data, question.prefix)
}
\arguments{
\item{data}{A dataframe.}

\item{question.prefix}{The common strings of that series of multiple response vaiables}
}
\description{
Frequence tabll of Multiple response variable .
}
