% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldruc.R
\name{mrtab1}
\alias{mrtab1}
\alias{mrtab}
\title{多选题分析函数.}
\usage{
mrtab(df, ..., select_value = 1, title_prefix = "多选题分析")

mrtab1(data, question.prefix)
}
\arguments{
\item{data}{A dataframe.}

\item{question.prefix}{The common strings of that series of multiple response vaiables}
}
\description{
多选题分析自定义函数。

进行多选题分析，类似stata中的mrtab
}
