% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldruc.R
\name{rec}
\alias{rec}
\title{recode variable using rec in sjmisc.}
\usage{
rec(
  x,
  ...,
  rec,
  as.num = TRUE,
  var.label = NULL,
  val.labels = NULL,
  append = TRUE,
  suffix = "_r",
  to.factor = !as.num
)
}
\description{
用来进行变量的重编码.映射规则比较简单。
}
